Atualização de scripts

+++ PostgreSQL/PROJECTS/003-DBWORK/SQL/create_DBWORK.sql
@@ -1,7 +1,7 @@
 
 CREATE TABLE tb_sector (
 	id BIGSERIAL NOT NULL UNIQUE PRIMARY KEY,
-	co_sector INTEGER NOT NULL UNIQUE,
+	co_sector BIGSERIAL NOT NULL UNIQUE,
 	nm_sector VARCHAR(30) NOT NULL,
 	ds_sector VARCHAR(300) NOT NULL,
 	tp_sector VARCHAR(30) NOT NULL,
@@ -11,7 +11,7 @@ CREATE TABLE tb_sector (
 
 CREATE TABLE tb_vacation (
 	id BIGSERIAL NOT NULL UNIQUE PRIMARY KEY,
-	co_vacation BIGINT NOT NULL UNIQUE,
+	co_vacation BIGSERIAL NOT NULL UNIQUE,
 	nm_vacation VARCHAR(30) NOT NULL,
 	ds_vacation VARCHAR(300) NOT NULL,
 	tp_vacation VARCHAR(30) NOT NULL,
@@ -22,7 +22,7 @@ CREATE TABLE tb_vacation (
 
 CREATE TABLE tb_frequency (
 	id BIGSERIAL NOT NULL UNIQUE PRIMARY KEY,
-	co_frequency BIGINT NOT NULL UNIQUE,
+	co_frequency BIGSERIAL NOT NULL UNIQUE,
 	nm_frequency VARCHAR(30) NOT NULL,
 	ds_frequency VARCHAR(300) NOT NULL,
 	tp_frequency VARCHAR(30) NOT NULL,
@@ -33,7 +33,7 @@ CREATE TABLE tb_frequency (
 
 CREATE TABLE tb_salary (
 	id BIGSERIAL NOT NULL UNIQUE PRIMARY KEY,
-	co_salary INTEGER NOT NULL UNIQUE,
+	co_salary BIGSERIAL NOT NULL UNIQUE,
 	nm_salary VARCHAR(30) NOT NULL,
 	ds_salary VARCHAR(300) NOT NULL,
 	tp_salary VARCHAR(30) NOT NULL,
@@ -44,7 +44,7 @@ CREATE TABLE tb_salary (
 
 CREATE TABLE tb_position (
 	id BIGSERIAL NOT NULL UNIQUE PRIMARY KEY,
-	co_position INTEGER NOT NULL UNIQUE,
+	co_position BIGSERIAL NOT NULL UNIQUE,
 	nm_position VARCHAR(30) NOT NULL,
 	ds_position VARCHAR(300) NOT NULL,
 	tp_position VARCHAR(30) NOT NULL,
@@ -54,7 +54,7 @@ CREATE TABLE tb_position (
 
 CREATE TABLE rl_position_user (
 	id_user BIGINT NOT NULL,
-	id_position INTEGER NOT NULL,
+	id_position BIGINT NOT NULL,
 	dt_initial CHAR(14) NOT NULL,
 	dt_final CHAR(14) NOT NULL,
 	
@@ -63,8 +63,8 @@ CREATE TABLE rl_position_user (
 
 
 CREATE TABLE rl_position_sector (
-	id_position INTEGER NOT NULL,
-	id_sector INTEGER NOT NULL,
+	id_position BIGINT NOT NULL,
+	id_sector BIGINT NOT NULL,
 	dt_initial CHAR(14) NOT NULL,
 	dt_final CHAR(14),
 	
@@ -74,7 +74,7 @@ CREATE TABLE rl_position_sector (
 
 
 CREATE TABLE rl_position_salary (
-	id_position INTEGER NOT NULL,
+	id_position BIGINT NOT NULL,
 	id_salary BIGINT NOT NULL,
 	dt_initial CHAR(14) NOT NULL,
 	dt_final CHAR(14),
@@ -85,7 +85,7 @@ CREATE TABLE rl_position_salary (
 
 
 CREATE TABLE rl_position_frequency (
-	id_position INTEGER NOT NULL,
+	id_position BIGINT NOT NULL,
 	id_frequency BIGINT NOT NULL,
 	dt_initial CHAR(14) NOT NULL,
 	dt_final CHAR(14),
@@ -96,7 +96,7 @@ CREATE TABLE rl_position_frequency (
 
 
 CREATE TABLE rl_position_vacation (
-	id_position INTEGER NOT NULL,
+	id_position BIGINT NOT NULL,
 	id_vacation BIGINT NOT NULL,
 	dt_initial CHAR(14) NOT NULL,
 	dt_final CHAR(14),
